+++ tests/common.rs
@@ -18,6 +18,8 @@ pub fn build_configuration() -> GooseConfiguration {
         log_file: "goose.log".to_string(),
         stats_log_file: "".to_string(),
         stats_log_format: "json".to_string(),
+        debug_log_file: "".to_string(),
+        debug_log_format: "json".to_string(),
         sticky_follow: false,
         manager: false,
         no_hash_check: false,